Tries to avoid copying termreps by "refreshing" a single "current" termrep cache.
Profiling 8Q indicated that the repeated copying of term reps and
creation of python objects is still a considerable task (10-20%).
This commit moves toward *not* copying a magnitude-free term to a
with-magnitude term when magnitudes are set.  Instead, the same
underlying termrep object is used.  This is fine, except that we
then need to keep track of the "current" path-set being used by
a model or TermForwardSimulator -- and we need to be able to
reset/refresh the magnitudes in any given term-rep-cache according
to a given model paramter vector.

This still needs more work, and probably means moving the
_interpolate_model_until_no_failures into the Model or fwdsim class.
It probably means creating a path-set object for better organization.

The main idea is that previously we specified a set of paths via a set of
per-circuit (pathmag-)thresholds and a "locked-in" set of high-magnitude-term-reps
(in a "repcache").  Now the repcaches for any given path-set share underlying
"memory" and so whenever a path-set is created it invalidates all the other
repcaches.  To "restore" an old repcache, we need to "refresh" it's repcache
so that all the (shared) underlying term reps (the "memory") is valid for
the parameter vector corresponding to the desired repcache.

It is now possible to perform this refresh using the 'refresh_magnitudes_in_repcache'
function, which is called following every call to _interpolate_model_until_no_failues.
The "select_paths_set" method only needs to be called when choosing a new
path set.  But by virtue of considering other path sets, when not-choosing
another path set we need to refresh the repcache of the current path set
so this repcache is still valid.