Reverts the experimental optimization of the last commit.
A bit more testing reveals that there's something not quite
correct about the last commit's update to avoid copying termreps.
The prior commit could still be useful later on if we want to
add this optimization properly, but for now we'll just stick
with copying termreps to make things more robust and intuitive.