Improves handling of the "unreliable2Q" gauge opt suite.
When "unreliable2Q" is specified as a gauge optimization suite
to do_stdpractice_gst it was handled differently based on whether
the target model's gauge group was trivial or not.  Namely, the
'single-2QUR' gauge optimization was *always* added in the trivial
gauge group case and it was only added when an unreliable 2-qubit
gate was present in all other cases (there's a default list of
unreliable 2Q gate names, or one can be provided via the advanced
options).  Furthermore, when no unreliable 2Q gates were recognized,
the 'single-2QUR' gauge optimization was just dropped silently.

This commit improves this situation by causing in all gauge-group
cases the 'single-2QUR' gauge opt to be added only when a 2Q gate is
recognized.  It also produces a warning when no 2Q gates are recognized,
so that the reason behind the omission of the "single-2QUR" option
is more apparent.