Adds more experimental code for pruned path calcs.
This commit adds the ability to compute analytic jacobians of
the "termgap penalty" introduced in the last commit, so this penalty
term can actually be used (i.e. it isn't too slow).  It remains to be
seen how useful this will be, but the idea is that by adding a penalty
term instead of a hard constraint the LM-optimization will be more
robust and will be less dependent on initial conditions.  This update
adds additional poly-rep methods to fastreplib.pyx that will need to
be added to slowreplib.py later on if we keep everything.

This commit also adds some new trial hacks in core.py in an attempt
to test on 2 qubits how the new penalty term works in practice.  Very
much a WIP.  For instance, a HACK in customlm.py has been added to
circumvent strange behavior whereby it seems difficult for the LM opt
to eliminate termgap failures by simply increasing the penalty prefactor.