Updates Circuit hashing & .tup to include line labels.
Previously (commit a909c9a89c164cc46ab64bf22557b6ef9f464d51) the
line labels of a Circuit were added to its hash and equality test.
This commit adds the line labels to the .tup attribute of a Circuit
and adds a .layertup property that returns just a tuple of the
layer labels (as .tup had done previously).  The line labels are
separated within the hashed tuple by a '@' sentiel, which avoids
having the labels within a nested tuple and better meshes with
Circuits without line labels ( ('*',) circuits), whose .tup does
not contain any '@' element.  Something to watch out for, and
which this commit patches in estimate.py, is that now tuple(circuit)
is *not* the same as circuit.tup, since the latter includes
the line labels, e.g., (('Gx',0),('Gy',1)) vs (('Gx',0),('Gy',1),'@',0,1).
The idea is that the .tup of a Circuit gives a complete version
of the Circuit as a tuple, and the approach here allows us to
continue to use just these tuples in lieu of the heavier Circuit
objects when working with core.py routines like do_mc2gst (a
performance gain).