Adds an initial OplessModel class.
This initial version duplicates the functionality in
errorratesmodel.py but is intended to more broadly be a Model
that is *not* an OpModel, i.e. it doesn't contain separately
parameterized operations which are used to construct circuit
probabilities.  The current version is pretty inefficient, and
contains some dummy members like .basis and .dim which are needed
by the framework but probably shouldn't be.