Continues experimental termgap objective function modifications.
Adds a "frozen hessian" option where we try to approximate the
computationally intensive termgap-penalty (probability gaps/error bars)
calculation with a simple quadratic based on an approximated
Hessian at the initial point.  This seems to be a bad idea, and
will likely be removed after this commit, as it takes a long time
to compute the approximate Hessian and the approximation isn't great.

A better path seems to be to move back to a simple non-worst-case
objective function (implemented in the new simple_* Chi2Function
members added in this commit).

This commit also adds more debugging structure for directly comparing
the path-integral-computed probabilities and gaps with the exact ones
to ensure that the exact values lie in the expected interval around
the approximate probabilities.