Adds custom-LM ability to check periodically for OOB condition.
The new `oob_check_interval` and `oob_action` arguments of custom_leastsq
allow the caller to give an objective function which takes an
optional `oob_check` argument that, when True, allows the objective
function to raise a ValueError (custom exception later?) when the
function is out-of-bounds.  This check can be performed at some
interval of outer iterations so it doesn't impact performance too
heavily.  The termgap chi2 objective function has been updated to
be of this type.

In core.py the _do_term_runopt function utilizes this mode of
 operation in an updated pruned-path-integral algorithm
whereby the optimization will stop and recompute paths as soon as
it tries to evaluate the objective function in an out-of-bounds
location, except on the final "sub"-iteration when it will just
reject out-of-bounds locations (since we won't be adding any more
paths).

This new algorithm does not require the evaluation of "termgap penalties"
so the compuation of polynomials used for the termgap jacobian have
been commented out within termevaltree.py.