Adds ability to stop termcalc-stage interations based on % allowed paths.
The find_minimal_paths_set(...) methods of TermForwardSimulator
and TermEvalTree now return the total number of paths in the
paths-set along with the maximum allowed number based on the fwdsim
parameters.  This allows _do_term_runopt(...) to exit the term-stage
loop early when the selected set of paths has over 90% of the
allowed paths (the thinking is that additional iterations won't
be able to add much more precision to the probabilities).