Updates termgap logl objective function to match latest chi2.
Now logl objective function for the termgap simtype is a
non-worst-case objective that checks for an out-of-bounds condition
when requested (similar to the chi2).  This commit also updates the
names of the "worst case" objective functions and jacobians by
prefixing them with "worstcase_" -- these functions aren't used
anymore and should probably be removed in the future.