Fixes a bug in custom LM routine.
Setting mu=0 initially in the custom_leastsq function was previously
non-problematic but became so when adding the out-of-bounds interval
behavior: if iteration 0 is deemed "converged" but with an out-of-bounds
interval > 1, the optimizer would then continue to restart from iteration
0's point but it would be iteration *1*, and mu would have been left
uninitialized.  This commit fixes this issue but 1) initializing mu to
1 instead of 0 so if this does happen at least the method will be able
to build up damping (if mu is 0 it will always remain 0!), and 2) initializing
mu using the init_munu at the beginning of the function and not just
within the "k==0" block.