Adds TermPathSet objects for cleaner "paths set" tracking.
Now a TermEvalTree holds a .pathset member containing information
about the paths it will use (or a fwdsim object will use) to
compute probabilities.  It's a little weird that the fwdsim object
builds a "SplitTreeTermPathSet" object which is never actually held
by any evaltree - the fwdsim treats all trees as "split" and operates
with the subtree's pathsets directly.  Maybe there's a nicer way to
do this.