Alters _do_term_runopt to expect a Model with already-prepped probs.
Previously, the initial model received by _do_term_runopt was not
expected to have an existing pathset, and one was selected at the
beginning of the function.  Now bulk_evaltree_from_resources calls
bulk_prep_probs (which selects a path set) and _do_term_runopt
assumes the model already has a selected path set.  This makes
transitioning between stages of a GST optimization smoother, and
avoids re-selecting a similar (or the same) set of paths when it
isn't needed.

Also - removes the bulk_prep_probs calls in the chi2 functions
added recently to allow those function to work - this isn't
needed anymore.