Adds ability to keep zero-count dataset lines and indicate "BAD" lines.
pygsti.io.load_dataset now accepts an 'ignoreZeroCountLines' argument,
that, when False, keeps a record of Circuits and aux-info corresponding
to rows with no data counts.

Secondly, the DataSet text file format has been expanded to
allow the user to specify "BAD" in place of the data counts for a
circuit.  This causes no (zero) counts to be read in for that line
or for the line to be ignored, depending on the value of
'ignoreZeroCountLines' as described above.  Warning messages from
load_dataset will be suppressed for "BAD"-lines, as this indicates
the user is already aware they contain no data.