Adds default time=0.0 args to Label __new__ methods so they can pickle.
I think anything inheriting from 'tuple' needs to have a __new__
method that takes a single argument so the tuple part of the object
can be properly initialized with un-pickling.  This commit does this
for all Label classes except CircuitLabel, which may have an issue
pickling but just isn't tested yet.