Fixes bug in MultiDataSet caused by not .copy()-ing index dicts.
Previously, a MultiDataSet would hold references to the circuit and
outcome-label index lists of the first DataSet added to it, essentially
assuming control over these dictionaries which could be updated later
on by the MultiDataSet when additional DataSets were added.  This broke
the original DataSet, whose index lists were manipulated underneath it.
This commit fixes this issue by having the MultiDataSet .copy() the
index dictionaries.