Adds explicit 'errcomp_type="gates"' in nqnoiseconstruction.py.
We believe that changing the error composition type won't really
matter when finding amplifying sequences, since to first order the
two error composition types are the same.  Still, it's nice to be
reminded we're using the "gates" error-composition-type for the
cloudnoise-sequence-generation functions in nqnoiseconstruction.py,
and this commit just makes that explicit.