Adds gauge-invariant-gateset representation table to standard report.
This report table is for testing an (experimental) way of representing
a gate set in a "mostly" gauge invariant way, as noted in the figure
caption (placed on the gauge-invariant-metrics-overview tab of the
report). Utilizes a brute-force numerical algorithm to find the desired
gate decomposition, which can probably be replaced with something more
clever (and faster) in the future.