Fixes a bug in Benchmarker.generate_success_or_fail_dataset.
When creating a success-fail DataSet, the collisionAction of the
created DataSet should (I think) be set to 'keepseparate' when
the standard datasets use this mode.  Furthermore (this was the
real bug), when iterating through the items of the 'standard'
datasets we needed to specify stripOccurrenceTags=True - otherwise
we could inject tagged data rows into the new DataSet.

This commit also cleans up the generate_success_or_fail_dataset
function a bit so it uses the row's .aux instead of looping through
the DataSet's .auxInfo.