Updates gauge-robust decomposition to ensure matrices are real.
Replaces 'kite_block_diag_*' functions (in matrixtools.py) with
'find_zero_communtant_connection', which works in the original
gate basis rather than an eigenbasis of G0.  This allows us to
ensure logarithms and the matrices in the final decomposition are
kept real, resulting in a *real* decomposition of G = F M G0 Finv.

This also fixes the issue of the iterative 'quick' method not
converging all the way, so it's likely that this function was doing
something wrong, perhaps assuming to work in a unique eigenbasis
of G0 when such a basis didn't exist.  All in all, the new method
is faster, cleaner, and most importantly ensure the F and M matrices
are real.