Updates report loading, replacing document.ready() handlers.
Now that an entire report is contained in a single html file,
the document.ready handlers that were scattered in the separate
table/plot/tab files, only to be called when needed, are all called
at the initial page load.  This created slow page loads, since all
of the plots (even on hidden tabs) were loaded up front.  This commit
causes that loading to be instead triggered by a 'load_loadable_item'
event, which the relevant plots and tables now add handlers for instead
of generating simple document.ready handlers.

This has the intended effect within a report, but won't work when the
tables/plots/switchboards are used in other contexts, such as within
an IPython notebook.  So this commit adds a "within_report" argument
to Switchboard.__init__ and a "within_report" render option for
WorkspaceOutput objects, that is set to True in factory.py and
merge_helpers.py when building a report (and defaults to False).  This
should allow the tables, plots, and switchboards to function within
notebooks as they had before, but this hasn't been tested and might
need some futher tweaking.