Fixes a bug in tree splitting (all trees updated).
When a tree was split, EvalTree._finish_split() would permute
the parent tree's circuits (to put them in a "nice" order so the
subtrees would just be *slices* of the parent) but wouldn't update
the other members of the tree relating to the elements' relationship
to the circuits, e.g. which effect/spam labels correspond to circuit k,
until after the create-subtree function of the derived class was called
multiple times to create the subtrees.  This was problematic since the
create_subtree functions would assume that the parent tree was in an
internally consistent state, i.e. the element-referring-members would
be in sync with the circuit list.

This commit fixes this issue, which would only surface when using
split trees (mulitple procs) *and* when the spam/effect labels
associated with different circuits is different (sparse data).  It
does this by adding a new EvalTree._update_element_indices method
that acts as a hook for the derived tree classes to update their
other (non-self[:]) members when self[:] is permuted.  This method
is called within _final_split *before* the subtrees are created.