Adds support for single-file jinja template merging.
Not tested yet, but the idea is that if the user specifies a .html
filename to create_standard_report (and other report gen functions)
a directory is *not* created and instead just the single prescribed .html
file is made, possible with syncing a sibling "offline" directory. In
this mode, link_to would have to be None (and btw link_to isn't currently
working again since the jinja update).