Adds simplify_circuits and todict methods to error rates models.
The simplify_circuits function is half-baked as it doesn't actually
return the first "raw circuit dict" that another Model would (it
returns None) - this doesn't seem to be needed though.  This allows
error rate models to be used in objective functions like logl, chi2,
etc.

The todict function returns the error rates of a model, which is
now different from the parameter vector of the model.