Fixes more bugs in WildcardBudget.update_probs.
The main bug fix is that we were locking p_k = f_k in "setC" of
update_probs, rather than p_k = q_k, which is what we should have
done and what is now done.  This commit also adds debug toggling
in the update_probs function, which should be removed in the future.