Adds robust loading of JS and CSS in reports.
Through the use of more complex jinja macros, the main.html jinja
template now inserts logic to attempt to load an online JS library
and, if this fails, to load it locally (if an offline/ directory
exists).  In the CSS case, we either insert the CSS directly (when
the sheet is *local*, i.e. has no online URL) or try to import
multiple redundant stylesheets (when an online URL is given).

The result is that report .html files may be used both apart from an
offline/ directory when there's an internet and vice versa, making them
more flexible.

Tne potential issue is that if the pygsti_dataviz.css can't be
inserted for some reason the code will fallback to inserting a
link to pygsti_dataviz.css.  This is mostly fine, but the image urls in
this CSS file will be broken: 'images/pyGSTi_icon.png' has been changed
to 'offline/images/pyGSTi_icon.png' anticipating that the sheet will
usually be inserted in the top-level html document, rather than being
import from the offline/ folder.  This seems like a very unlikely and
minor issue even if it does occur.