Speeds up and further fixes wildcard budget optimization.
Adds a wildcard ObjectiveFunction object to reduce
the amount of computation needed during the wildcard budget
optimization.  This still isn't super fast, and should be profiled,
but it's getting better.  This fixes some minor bugs (like handling
the f_k=0 terms properly) and updates the optimization algorithm
so that once a small enough eta that the misfit-penalty is
zero (the solution meets the desired goodness-of-fit constraints)
is found, the optimzation converges.  (After fixing the objective
function in prior commits, it was seen that the complex method of
trying to find the "shore of the island" was not needed.)

This commit also adds a SPAM-budget element to PrimitiveOpsWildcardBudget
objects, controllable by an add_spam argument.  The spam budget is
not currently displayed anywhere in reports.