Adds ability to set the background color of color box plots.
The 'colorboxplot_bgcolor' advanced option to report creation
function now controls the background of color box plots.  This
update required removing masked-arrays from the lin-log colorscale's
normalize() method, which for some reason prevents the plotly background
color from having effect.  This update *may break* the latex generation
of these plots - we need to check this and fix it if needed.