Simplifies start_budget behavior.
Previously, the start_budget argument to PrimitiveOpsWildcardBudget
added some "noise" on top of the given value as this seemed to
assist the optimizer in finding a good final budget.  This, however,
doesn't seem necessary now that the objective function has been
fixed, and just adds complication where none is needed.  So now
upon initialization all the elements of a PrimitiveOpsWildcardBudget
object are set exactly to start_budget.