Fixes jsFallbackFromCDN(...) and how it is called in HTML report templates.
Previously jsFallbackFromCDN was called as an "onerror" handler for
CDN-loading <script> tags in our report main.html templates.  This doesn't
work correctly when the script fails to load, at least in some browsers,
since the onerror event is triggered asynchronously and by the time
jsFallbackFromCDN gets called it's too late to load the needed library.

This commit updates the jinja macro so that each CDN-loading script tag
is followed by a call to jsFallbackFromCDN, which tests if the desired
library has indeed been loaded and loads the local version if it hasn't.

Within jsFallbackFromCDN, document.write(...) is used instead of the
nicer approach of creating DOM elements as objects because we need
to insert the new script tag *immediately* (not just at the end of
document.head), and it seemed more confusing/ugly to add the JS code needed
to "find the current script block", so a DOM element could be inserted
correctly using the object-oriented approach.