IPython notebook display elements are imported in `Workspace` as-needed.
These imports are somewhat expensive (~0.15s on my machine) and are
only used in the special case when running in an IPython notebook. In
principle, this change may introduce marginal overhead when rendering
workspace objects in a notebook; in my own tests I haven't noticed any
slow-down.

Additionally, `in_ipython_notebook` is replaced with a function of the
same name which checks the actual runtime environment. Note that the
previous try/except method of determining if we were running in a
notebook was incorrect, giving a false-positive if the user simply had
IPython notebook installed.