Makes updates to temporarily disable RB imports and fix unit tests.
Unit tests have been updated to sync with the Circuit changes, namely
that Circuits now hash with their line-labels.

Since the RB code is currently in a known-broken state, this commit
comments out the "from . import rb" from extras/__init__.py, and
skips all of the RB unit tests (to be re-enabled when the RB code
is up and running again).