Adds new GaugeRobustMetricTable and underlying infrastructure.
This new table attempts cast the errors in a gate set as being either
"intrinsic" to a gate - i.e. in the eigenvalues - or "relational" -
i.e. in the relative eigenvectors.  Standard metrics like the diamond
norm and infidelity are used to compute the distance between a gate
in its "best case" gauge and the target, and between the target gates
and themselves in another gate's "best case" gauge.  This commit adds
the framework for this by factoring `get_a_best_case_gauge_transform`
out of `project_to_target_eigenspace` in optools.py, and this method
underlies the values computed in the new report table.

The table doesn't seem to be working right yet, but it does show
up nicely in reports.  This is likely due to the fact that the best-case
gauge needs to be further optimized by taking into account the other
gates.