Fixes bugs in gauge-robust metrics computation, and adds additional gauge opt.
Several bugs are fixed in the way gauge-robust metrics are computed,
and an additional numerical gauge-optimization step is added to the
"best case gauge" returned by the optools.py function.  The bug fixes
resulted in the procedure "mostly" working for test cases: in a
CPTP fit of 1Q data the results indicated a substantial relational
error between two gates (Gx and Gy) even when there shouldn't have been
one (data was generated with just depolarizing noise) and this was
largely remedied by the additional gauge optimization, which tweaks
the best-case gauge of a gate to also consider the other gates at
a much lower weighting.