Adds some commented-out debug statements for infidelity behavior in GaugeRobustMetricTable.
The infidelity seems to behave badly in the gauge-robust metric
table (namely, it shows significant negative values for relational
error when they there should be zero relational error - and is, as
reported by real metrics like the diamond or frobenius norm).  We're
not sure if this is just an artifact of the infidelity or a bug
somewhere.