Fixes issues with running GST after Circuit-hashing updates.
In the past many of the core.py routines would prefer to work with
tuples of labels instead of Circuit objects, as the former are
lighter and simpler.  This wasn't a problem, since Circuits where
hashed (and tested for equivalence) just like their underlying tuple
of layer labels.  With the recent update to Circuit, which includes
the line-labels in a Circuit's hash, this functionality was broken.

This commit comments out and marks w/"TODO REMOVE" several blocks
in core.py that would convert circuit lists to tuple lists for
increased processing speed.  It also adds several more if/else blocks
around calls to find_replace_tuple_list(...) so that this function
(which generates tuples and would therefore break the process flow
now) is not called unless opLabelAliases is not None.  In the future
the use of op/gate-label aliases should either be fixed by creating
a Circuit-compatible replacement function or removed entirely - I
don't think aliases are currently used anywhere (and they're broken
now, in any case).

Now running GST seems to work again, now using line-label-aware
Circuit hashing.