Fixes a bug in path-integral calculations causing negative gaps.
Stress testing on more qubits and longer sequences revealed a bug
in the path-integral calculations: the .compact_coeff member of
SVTermRep (added somewhat recently for performance) was not "synced"
with the .coeff member after a call to SVTermRep.mapvec_indices_inplace.
That is, the .compact_coeff member still had *local* parameter indices
in it whereas the .coeff (correctly) had *global* indices.  This
commit fixes this issue, leaves several residual debug statements for
cleanup in the next commit, and updates the (gaps >= -1e-6) assertion
to (gaps >= 0), since we no longer expect any tolerance to be needed
(see below).

This commit also updates the algorithm for counting/summing paths
within fastreplib.pyx.  The count_paths(...) function has been
updated so that while the SMALL variable is used to prevent the
current magnitude from being zero (this is bad, since if the magnitude
is zero we can't recover what anything from it - it's as if it loses
all memory of where we're at!), the new 'current_nzeros' argument
tracks how many 0s there are in the path product.  If there are any,
then we know that the product is actually 0, not whatever residual
nonzero value is left from having multiplied SMALL through.  This is
why we don't need to allow a tolerance in the gaps >= 0 condition. The
value of SMALL is still 1e-10, but it shouldn't need to be actually
small anymore (and this may help with roundoff errors) - we should
test this in the future.

The path counting/traversing in slowreplib.py still needs to be
synchronized with the more updated version in fastreplib.pyx.