Updates SMALL from 1e-10 to 1e-5.  Checks that this doesn't change results.
Test on simple 1Q example.  Now that current_nzeros accounts for
zeros within the product of term magnitudes that is the path magnitude,
the value of SMALL doesn't matter.  This test confirms that expectation.
Setting a larger value of SMALL might (?) reduce the amount of roundoff
error incurred when multiple SMALL values are included in the current_mag
product.