Updates gauge-robust calculations, making them more robust (haha).
Updates optools.py's get_a_best_case_gauge_transform function to
use a more robust (but less tested!) calculation method and to
return more of its intermediate values for use in the gauge-robust
decomposition table.  In the future, we should make the functions
now defined within get_a_best_case_gauge_transform(...) into top-level
functions, and cleanup the now-unused old way of computing a best
gauge transform based on projecting onto the estimates eigenvectors
onto the eigenspaces of the target (the new method instead replaces
the eigenvalues of the estimate of those with the target and processes
the estimate and target as having the same possibly (probably!)
degenerate eigenvalues).

This commit also updates the gauge-robust tables to be resilient to
bad cases when the target and estimate are far apart - in cases
where, for instance, a real logarithm cannot be taken the tables
will now show blanks.

Finally, this commit makes some minor updates/fixes ton matrixtools.py,
related to the get_a_best_case_gauge_transform function.