Finishes implementation of the embed_figures report (advanced) option.
Adds plumbing so that when embed_figures=False the loadable-item
handlers are *not* created (since we don't use loadable-item triggering
when we don't embed figures; the figures just initialize themselves
immediately once they're loaded).  This is done via the 'use_loadable_items'
WorkspaceOutput option and Switchboard argument, which are connected
to the embed_figures argument in factory.py and merge_helpers.py.

The resulting embed_figures=False reports now display correctly when
served from a webserver or appropriate security allowances are made
in the user's web browser.  When AJAX calls fail there is no longer
an splash screen informing the user of this -- maybe we should add
that back.