Initial commit on feature-protocols branch.  Initial GST protocol objects.
Begins implementation of new API-layer that uses "protocol objects",
which thus far entail "input", "data", and "results" objects in
addition to the central "protocol" objects.  The interworking of
these objects is still in flux, as this is a WIP.  This commit
defines some framework in protocols/protocol.py and roughly implements the
GST functionality in drivers/longsequence.py in protocols/gst.py and
protocols/modeltest.py.  A number of things still need to be done
on the GST front, including:
- once the new objects are working, switch the drivers/longsequence.py
  routines to use them.
- reconile the gauge optimization & packaging code in the GST and
  StandardPracticeGST protocols (e.g. model-test uses same packaging in
  longsequence.py, but separate objects now; gauge opt suites vs param-dict).
- figure out what to do with the Results object (how this fits into
  the code structure).