Adds initial protocol objects for RB.  Still needs work.
A base-class benchmarking protocol creates result "quantities"
(now nested dicts) from single or multi-pass data, and this serves
as the foundation for other protocols (currently just VolumetricBenchmark).
The thinking is that using a to-be-created MultiProtocol (?) object
one will be able to combine protocols that act on the same input
in such a way that they can cache results for one another in a shared
results object.