Another update to "protocol objects", still a WIP.
Adds "Simultaneous" and "Multi" input and protocol objects, for
creating hierarchies of protocols.  Simultaneous inputs are those
performed in parallel on disjoint sets of qubits; Multi-inputs
specify multiple input structures/protocols on a single data set.

Adds a "cache" member to ProtocolData, which is intended to hold
derived data that *only* depends on the input specified within
the ProtocolData and does not depend on additional protocol
parameters (e.g. different GST estimates should *not* be stored
here, but marginalized data-sets based on the 'qubit_labels' of
an input are fair game).

Adds minor fixes to get GST and RB/VB protocols working so some
simple testing can be done.