Fixes bug in TensorProdSPAMVec causing bad probability calcs.
An assignment in _update_rep() to self._rep.base instead of
self._rep.base[:] caused the underlying representation object to
have a memory pointer that could be (and sometimes was) garbage-
collected!  When this happened all the probabilities computed using
the spam vector were, somewhat ironically, garbage.  This commit
fixes this issue and adds some additional code to try to ensure
that the numpy arrays passed to DMStateRep "own" their own data.
This doesn't seem to be essential, but it seems like it can only
help prevent unwanted garbage collection.