Add "sub-data" to protocol-object paradigm.
As a step toward being able to nicely store the protocols objects
persistently, this commit adds the concept of having sub-ProtocolData
objects within a ProtocolData object which mirror the sub-inputs in
its input.  Whenever a ProtocolData object has a MultiInput input,
it must also store sub-data elements.  It is the responsibility of
a MultiInput object to create sub-ProtocolData objects, as these
input objects contain all the information needed to take a "parent"
dataset and generate a sub-data element from it.

This commit also has derived Protocol classes implement _run rather
than run, allowing run(...) to implicitly create & add MultiProtocol
objects when a Protocol is run(...) on data containing multiple inputs
and data.