Updates ProtocolResults object and way of composing results.
Since it seems useful to have specialized ProtocolResults objects
that have some knowledge of their contents (e.g. ModelEstimateResults),
MultiProtocol now stores the entire sub-results objects within
the self.qtys of the parent ProtocolResults object.

To avoid having a mess of nested Results that are hard to navigate,
the ProtocolResults class now exposes self.qtys as a dictionary.
Also, the added NamedDict class along with asdict and asdataframe
(a still-in-progress functionality to export to a pandas dataframe)
ProtocolResults methods will facilitate exploration of a results object.