Factors out "multi-pass" element of RB protocols to MultiPassProtocol.
Since we should be able to aggregate easily using dataframes derived
from ProtocolResult objects, we don't need to work so hard to make
the "pass" index the innermost one.  Furthermore, the ability to consider
a protocol on multiple passes corresponding to the different elements
of a MultiDataSet seems like a generically useful thing.

This commit factors the support for multiple passes out of the VB
code (simplfying it) and creates the MultiPassProtocol class, which
performs a single protocol and input on multiple datasets (the
members of a MultiDataSet).  The ProtocolData object, when it contains
a MultiDataSet, "knows" that it contains passes, and in this case
its ._subdatas member contains a dictionary of the ProtocolData
objects corresponding to the passes (this trumps _subdatas holding
data objects for multiple inputs, as each single-pass data object
can be a multi-input object).

Finally, implicit support for MultiPassProtocol has been added,
so when a protocol is run on multi-pass data it automatically
gets its _run called on each single-pass separately (to aggregate,
we must either aggregate the MultiDataSet beforehand or create a
protocol that implements run(...)).