Adds IO (load/write) functions for protocol objects; revamps path filter.
Adds to from_dir and write methods to ProtocolInput, MultiInput,
and ProtocolData (still no IO for results objects).  Untested.

Adds filter_paths, view, and get_tree_paths functions to ProtocolInput
so that filtering can be performed at the input "level".  This
improves the code structure and flexibility regarding the filtering
of multi-input/data objects, which hopefully will be useful down
the line.