Fixes a bug in Circuit.str behavior when circuit is empty.
When obtaining the string for an empty circuit, "{}" was used even
when the Circuit had non-trivial (not () or ('*',) ) .line_labels.
In this case, reading the circuit back in loses the line labels,
which is bad.  This commit fixes this bug, so that now, e.g.
"{}@(Q0,Q1)" is the string representation for an empty circuit on
qubits Q0 and Q1.