Adds IO for ProtocolResults objects, along with some structure adjustments.
To be more similar with the Input-to-Data relationship, and because
is seems more inuitive, a ProtocolResults now *holds* a ProtocolData
object rather than being one.  Results objects have 'name' attributes,
usually, as do Protocol objects now, which typically label the
protocol instance they contain results for.  ProtocolResults objects
hold data in two places:
- the ._subresults dict, which holds references to results objects
  for sub-data.
- the .qtys dict, which holds everything else.

It seemed useful to distinguish these types, as the subresults data
should be (and is) stored in the appropriate sub-directory, whereas
the qtys are properly the possession of the top-most results object
and are stored under the top-level  results/<results-name>/ directory.

This distinction is blurred when accessing a ProtocolResults object,
as its dictionary methods lump together the .qtys and ._subresults
dicts.

It's still somewhat unclear how to deal with a complex protocol
like VolumetricBenchmarkGrid, which essentially aggregates the
results of an underlying protocol.  Perhaps breaking it into a
post-processor object that acts on an existing ProtocolResults
object and expects certain quantites to be computed is the best
way to allow this flexibility.