Addition of (initial) default-protocol support.
This commit adds the ability for input objects to hold default
Protocol objects (and serialize them now!) so that once a data object
is created a set of default protocols can be run on it.  A missing
component is how Multi-Protocols are re-initialized, and ProtocolResults
now utilizes self.auxfile_types in two different ways.  A WIP.