Updates protocol object: results write a meta.json and adds category to inputs.
The ResultsProtocolDir class now writes a meta.json with its class
name in the root/results directory, simliar to the behavior of
the other tree nodes.  This allows us to remove the special if-metadir-is-none
processing in TreeNode._init_children, and may add future-proofing.

The _childcategory is added to the TreeNode class, so that all treenodes
can have a category for labelling their children.  This anticipates
a method for casting a tree into a dataframe with column headers, as
these 'categories' roughly correspond to column names.