Adds PredictedVolumetricBenchmark protocol.
Adds a custom_data_src argument to the VolumetricBenchmark protocol
which causes it to use, instead of the data derived from the passed-in
the ProtocolData object, the success/fail data found in the custom_data_src
argument (if it's a dict) or the success/fail data generated by
custom_data_src (if it's a model).  The PredictedVolumetricBenchmark
derived class makes it clear how to invoke VolumetricBenchmark using
this argument.