Fixes bug in derivative of "D"-type (depolarizing-only) models.
There was an error introduced when "upgrading" the einsum lines in
LindbladErrorGen's _dOdp(...) method, replacing einsums with np.transpose
calls that didn't sum on one of the axes.  This resulted in a derivative
that was the wrong shape, raising a numpy error.  This commit fixes the
issue with the derivative calculation and should also fix the Hessian
calculation, though this wasn't tested.